requirements.txt
# local package
-e .

# external requirements
click==8.1.7
Sphinx==7.2.6
flake8==7.0.0
python-dotenv>=0.5.1

# Librerie di Deep Learning
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# TensorFlow con supporto GPU
# NOTA: CUDA Toolkit 11.8 e cuDNN 8.6 devono essere installati separatamente dal sistema
# Vedere istruzioni di installazione nel CUDA_SETUP.md
tensorflow>=2.13.0
tensorflow-datasets>=4.9.0

# Strumenti di sviluppo
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.5.0

# Google Colab compatibility
# NOTA: google-colab è disponibile solo nell'ambiente Colab, non installabile localmente
# Le seguenti dipendenze supportano l'integrazione Colab-locale
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# Additional useful libraries for data science
scipy>=1.10.0
plotly>=5.0.0
tqdm>=4.60.0
pillow>=9.0.0
opencv-python>=4.8.0

# Version checking and package management
packaging>=21.0

# Git integration for Colab sync
GitPython>=3.1.0
